Update profile_zernike.ipynb
'Fixed bugs, ran a first tests and it was not faster than the method implemented right now. The columns of the output are not in the same format of the output from zernike_radial but all the required computations are executed.

Changing the format should just require for loops in the last if's of the function'